+++ gala/potential/common.py
@@ -1,14 +1,10 @@
-# Standard library
 import inspect
 
-# Third-party
-import astropy.units as u
-from astropy.utils import isiterable
 import numpy as np
+from astropy.utils import isiterable
 
-# Project
-from ..util import atleast_2d
-from ..units import UnitSystem, DimensionlessUnitSystem
+from gala.units import DimensionlessUnitSystem, UnitSystem
+from gala.util import atleast_2d
 
 
 class PotentialParameter:
@@ -21,24 +17,28 @@ class PotentialParameter:
     physical_type : str (optional)
         The physical type (as defined by `astropy.units`) of the expected
         physical units that this parameter is in. For example, "mass" for a mass
-        parameter.
+        parameter. Pass `None` if the parameter is not meant to be a Quantity (e.g.,
+        string or integer values).
     default : numeric, str, array (optional)
         The default value of the parameter.
     equivalencies : `astropy.units.equivalencies.Equivalency` (optional)
         Any equivalencies required for the parameter.
     """
 
-    def __init__(self, name, physical_type="dimensionless", default=None,
-                 equivalencies=None):
+    def __init__(
+        self, name, physical_type="dimensionless", default=None, equivalencies=None
+    ):
         # TODO: could add a "shape" argument?
         # TODO: need better sanitization and validation here
 
         self.name = str(name)
-        self.physical_type = str(physical_type)
+        self.physical_type = str(physical_type) if physical_type is not None else None
         self.default = default
         self.equivalencies = equivalencies
 
     def __repr__(self):
+        if self.physical_type is None:
+            return f"<PotentialParameter: {self.name}>"
         return f"<PotentialParameter: {self.name} [{self.physical_type}]>"
 
 
@@ -54,7 +54,7 @@ class CommonBase:
         sig_parameters = []
 
         # Also allow passing parameters in to subclassing:
-        subcls_params = kwargs.pop('parameters', {})
+        subcls_params = kwargs.pop("parameters", {})
         subcls_params.update(cls.__dict__)
 
         for k, v in subcls_params.items():
@@ -68,11 +68,14 @@ class CommonBase:
             else:
                 default = v.default
 
-            sig_parameters.append(inspect.Parameter(
-                k, inspect.Parameter.POSITIONAL_OR_KEYWORD, default=default))
+            sig_parameters.append(
+                inspect.Parameter(
+                    k, inspect.Parameter.POSITIONAL_OR_KEYWORD, default=default
+                )
+            )
 
         for k, param in sig.parameters.items():
-            if k == 'self' or param.kind == param.VAR_POSITIONAL:
+            if k == "self" or param.kind == param.VAR_POSITIONAL:
                 continue
             sig_parameters.append(param)
         sig_parameters = sorted(sig_parameters, key=lambda x: int(x.kind))
@@ -105,7 +108,8 @@ class CommonBase:
                 f"{self.__class__.__name__}: Potential and Frame classes only "
                 "accept parameters as positional arguments, all other "
                 "arguments (e.g., units) must now be passed in as keyword "
-                "argument.")
+                "argument."
+            )
 
         parameter_values = dict()
 
@@ -123,8 +127,10 @@ class CommonBase:
             parameter_values[k] = val
 
         if len(kwargs):
-            raise ValueError(f"{self.__class__} received unexpected keyword "
-                             f"argument(s): {list(kwargs.keys())}")
+            raise ValueError(
+                f"{self.__class__} received unexpected keyword "
+                f"argument(s): {list(kwargs.keys())}"
+            )
 
         return parameter_values
 
@@ -134,19 +140,25 @@ class CommonBase:
         pars = dict()
         for k, v in parameters.items():
             expected_ptype = cls._parameters[k].physical_type
-            expected_unit = units[expected_ptype]
+            expected_unit = (
+                units[expected_ptype] if expected_ptype is not None else None
+            )
             equiv = cls._parameters[k].equivalencies
 
-            if hasattr(v, 'unit'):
-                if (not isinstance(units, DimensionlessUnitSystem) and
-                        not v.unit.is_equivalent(expected_unit, equiv)):
-                    msg = (f"Parameter {k} has physical type "
-                           f"'{v.unit.physical_type}', but we expected a "
-                           f"physical type '{expected_ptype}'")
+            if hasattr(v, "unit"):
+                if not isinstance(
+                    units, DimensionlessUnitSystem
+                ) and not v.unit.is_equivalent(expected_unit, equiv):
+                    msg = (
+                        f"Parameter {k} has physical type "
+                        f"'{v.unit.physical_type}', but we expected a "
+                        f"physical type '{expected_ptype}'"
+                    )
                     if equiv is not None:
-                        msg = (msg +
-                               f" or something equivalent via the {equiv} "
-                               "equivalency.")
+                        msg = (
+                            msg + f" or something equivalent via the {equiv} "
+                            "equivalency."
+                        )
 
                     raise ValueError(msg)
 
@@ -156,27 +168,28 @@ class CommonBase:
                 if v.unit.physical_type != expected_ptype:
                     v = v.to(expected_unit, equiv)
 
+                v = v.decompose(units)
+
             elif expected_ptype is not None:
                 # this is false for empty ptype: treat empty string as u.one
                 # (i.e. this goes to the else clause)
 
                 # TODO: remove when fix potentials that ask for scale velocity!
-                if expected_ptype == 'speed':
-                    v = v * units['length'] / units['time']
+                if expected_ptype == "speed":
+                    v = v * units["length"] / units["time"]
                 else:
                     v = v * units[expected_ptype]
 
-            else:
-                v = v * u.one
+                v = v.decompose(units)
 
-            pars[k] = v.decompose(units)
+            pars[k] = v
 
         return pars
 
     def _remove_units_prepare_shape(self, x):
         from gala.dynamics import PhaseSpacePosition
 
-        if hasattr(x, 'unit'):
+        if hasattr(x, "unit"):
             x = x.decompose(self.units).value
 
         elif isinstance(x, PhaseSpacePosition):
@@ -197,7 +210,7 @@ class CommonBase:
         """
         Make sure that t is a 1D array and compatible with the C position array.
         """
-        if hasattr(t, 'unit'):
+        if hasattr(t, "unit"):
             t = t.decompose(self.units).value
 
         if not isiterable(t):
@@ -207,21 +220,28 @@ class CommonBase:
 
         if len(t) > 1:
             if len(t) != pos_c.shape[0]:
-                raise ValueError("If passing in an array of times, it must have a shape "
-                                 "compatible with the input position(s).")
+                raise ValueError(
+                    "If passing in an array of times, it must have a shape "
+                    "compatible with the input position(s)."
+                )
 
         return t
 
     # For comparison operations
     def __eq__(self, other):
-        if other is None or not hasattr(other, 'parameters'):
+        if other is None or not hasattr(other, "parameters"):
             return False
 
         # the funkiness in the below is in case there are array parameters:
         par_bool = [
             (k1 == k2) and np.all(self.parameters[k1] == other.parameters[k2])
-            for k1, k2 in zip(self.parameters.keys(), other.parameters.keys())]
-        return np.all(par_bool) and (str(self) == str(other)) and (self.units == other.units)
+            for k1, k2 in zip(self.parameters.keys(), other.parameters.keys())
+        ]
+        return (
+            np.all(par_bool)
+            and (str(self) == str(other))
+            and (self.units == other.units)
+        )
 
     # String representations:
     def __repr__(self):
@@ -229,10 +249,10 @@ class CommonBase:
 
         keys = self.parameters.keys()
         for k in keys:
-            v = self.parameters[k].value
+            v = self.parameters[k]
             post = ""
 
-            if hasattr(v, 'unit'):
+            if hasattr(v, "unit"):
                 post = f" {v.unit}"
                 v = v.value